testutils: move ListenerRegistry to leaf package
We'll want to access it from `base`, which can't import `testutils` due to
dependency cycles.
