+++ pkg/testutils/BUILD.bazel
@@ -9,7 +9,6 @@ go_library(
         "files.go",
         "hook.go",
         "keys.go",
-        "listener.go",
         "net.go",
         "pprof.go",
         "soon.go",
@@ -34,7 +33,6 @@ go_library(
         "//pkg/util/syncutil",
         "//pkg/util/timeutil",
         "@com_github_cockroachdb_errors//:errors",
-        "@com_github_stretchr_testify//require",
     ],
 )
