+++ pkg/testutils/listenerutil/BUILD.bazel
@@ -1,0 +1,13 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "listenerutil",
+    srcs = ["listener.go"],
+    importpath = "github.com/cockroachdb/cockroach/pkg/testutils/listenerutil",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/util/syncutil",
+        "@com_github_cockroachdb_errors//:errors",
+        "@com_github_stretchr_testify//require",
+    ],
+)
