+++ pkg/testutils/listenerutil/listener.go
@@ -8,7 +8,7 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-package testutils
+package listenerutil
 
 import (
 	"net"
